ProjectRevision constructor needs long; implement first and last revision information for the commit log.

+++ alitheia/tds/src/eu/sqooss/service/tds/ProjectRevision.java
@@ -141,7 +141,7 @@ public class ProjectRevision {
      * There is no date associated with this until the project revision
      * is applied to a specific SVN repository.
      */
-    public ProjectRevision(int revision) {
+    public ProjectRevision(long revision) {
         this();
         this.revision = revision;
         haveRevision = true;
